Reset tournament accounts (#720)
